count updated with some other modificaions

+++ TimeSheetManagement/employeeskill_service/src/main/java/com/feuji/employeeskillservice/bean/EmployeeBean.java
@@ -15,7 +15,7 @@ import lombok.ToString;
 @ToString
 public class EmployeeBean {
 
-	private Long employeeId;
+	private Integer employeeId;
 	private String employeeCode;
 	private String firstName;
 	private String middleName;
